requirements.txt
# AI Training Program - Python Dependencies
# =========================================
# Install with: pip install -r requirements.txt

# Core LLM Libraries
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.3.0

# Web Frameworks
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3

# Vector Database & Embeddings
chromadb>=0.4.22
sentence-transformers>=2.3.0

# Data Processing
tiktoken>=0.5.2
numpy>=1.24.0
pandas>=2.0.0

# HTTP & Async
httpx>=0.26.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# Database
aiosqlite>=0.19.0
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tenacity>=8.2.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.1.0
ruff>=0.1.0

# Optional: LangChain ecosystem (uncomment if needed)
# langchain>=0.1.0
# langchain-anthropic>=0.1.0
# langchain-openai>=0.0.5
# langgraph>=0.0.20
